Refatorando testes de promotion acrescentando o atributo name ao user utilizado para login. Utilizando do setup para replicar o mesmo user a cada suit de test. Criando um arquivo para os testes em que não há usuário logado.
