+++ test/system/promotions/promotion_show_test.rb
@@ -4,12 +4,15 @@ class PromotionShowTest < ApplicationSystemTestCase
 
   include LoginMacros
 
+  def setup
+    @user = login_user
+  end
+
   test 'should search field in the show promotion' do
-    user = login_user
 
     promotion = Promotion.create!(name: 'Natal', description: 'Promoção de Natal',
                           code: 'NATAL10', discount_rate: 10, coupon_quantity: 100,
-                          expiration_date: '22/12/2033', user: user)
+                          expiration_date: Time.zone.tomorrow, user: @user)
 
     promotion.generated_coupons!
 
@@ -21,11 +24,10 @@ class PromotionShowTest < ApplicationSystemTestCase
   end
 
   test 'hidden research field without coupons' do
-    user = login_user
 
     promotion = Promotion.create!(name: 'Natal', description: 'Promoção de Natal',
                       code: 'NATAL10', discount_rate: 10, coupon_quantity: 100,
-                      expiration_date: '22/12/2033', user: user)
+                      expiration_date: Time.zone.tomorrow, user: @user)
 
 
     visit promotion_path(promotion)
@@ -33,26 +35,11 @@ class PromotionShowTest < ApplicationSystemTestCase
     assert_no_selector "div#search_coupon"
   end
 
-  test 'can not search a coupon without login' do
-    user = User.create!(email: 'mclovin@iugu.com.br', password: '1234567')
-
-    promotion_natal = Promotion.create!(name: 'Natal', description: 'Promoção de Natal',
-                                        code: 'NATAL10', discount_rate: 10,
-                                        coupon_quantity: 100, expiration_date: '22/12/2033', user: user)
-
-    promotion_natal.generated_coupons!
-
-    visit search_coupon_promotion_path(promotion_natal, query: 'NATAL10-0099')
-
-    assert_current_path new_user_session_path
-  end
-
   test 'search a coupon by exact code' do
-    user = login_user
 
     promotion = Promotion.create!(name: 'Natal', description: 'Promoção de Natal',
                               code: 'NATAL10', discount_rate: 10, coupon_quantity: 100,
-                              expiration_date: '22/12/2033', user: user)
+                              expiration_date: Time.zone.tomorrow, user: @user)
 
     promotion.generated_coupons!
 
@@ -72,11 +59,10 @@ class PromotionShowTest < ApplicationSystemTestCase
   end
 
   test 'search a coupon by invalid code' do
-    user = login_user
 
     promotion = Promotion.create!(name: 'Natal', description: 'Promoção de Natal',
                               code: 'NATAL10', discount_rate: 10, coupon_quantity: 100,
-                              expiration_date: '22/12/2033', user: user)
+                              expiration_date: Time.zone.tomorrow, user: @user)
 
     promotion.generated_coupons!
 
@@ -91,11 +77,10 @@ class PromotionShowTest < ApplicationSystemTestCase
   end
 
   test 'search can not find coupon' do
-    user = login_user
 
     promotion = Promotion.create!(name: 'Natal', description: 'Promoção de Natal',
                               code: 'NATAL10', discount_rate: 10, coupon_quantity: 100,
-                              expiration_date: '22/12/2033', user: user)
+                              expiration_date: Time.zone.tomorrow, user: @user)
 
     promotion.generated_coupons!
